Create code.
This code uses setInterval() to switch the colors of the traffic light every 3 seconds. The colors are represented using the backgroundColor property of the bulb elements, which is set to red, yellow, green, or gray (to represent the off state). The if statements check which color is currently displayed and switch to the next one in the sequence.